extract functions to util cookie.ts
