--- packages/frontend/src/hooks/useCookies/canStoreCookies.ts
@@ -1,8 +1,0 @@
-import { CookieNames } from "."
-import getCookie from "./getCookie"
-
-const canStoreCookies = () => {
-    return getCookie(CookieNames.Consent) == 'true' ? true : false 
-}
-
-export default canStoreCookies
