--- packages/frontend/src/hooks/useCookies/getCookie.ts
@@ -1,7 +1,0 @@
-import Cookies from "js-cookie"
-
-const getCookie = (cookieName: string,) => {
-    return Cookies.get(cookieName)
-}
-
-export default getCookie
