--- packages/frontend/src/hooks/useCookies/index.ts
@@ -1,3 +1,0 @@
-export enum CookieNames {
-    Consent = 'SqCookieControl',
-  }
